fix(eql): widen wire max mana to uint32
m_wireMaxMana was uint16_t while the eql stat-sync wide form carries int64
values and setMana() takes uint32 — the same truncation class as the spell-id
widening. Widen the member, the wireMaxMana() accessor, and the maxMana local
in protoencoder that would otherwise re-narrow it. Not serialized by
savePlayerState, so no magicStr bump. m_mana/m_maxMana stay uint16
deliberately: they are stock cross-target fields.

Also correct an eqldispatch comment claiming the wide form "plays the role of
Live's absent OP_ManaChange" — OP_ManaChange does ride the Legends wire
(0x2255, S>C 20B); it just fires only on a cast and carries no max.

+++ src/player.h
@@ -133,7 +133,7 @@ public:
    // Exact, gear+buff-inclusive max mana from the eql stat-sync wide form
    // (OP_HPUpdate 0xa5c0) — the server's own value, so it tracks equip/unequip
    // and buffs. 0 on Live, which never sends max mana.
-   uint16_t wireMaxMana() const { return m_wireMaxMana; }
+   uint32_t wireMaxMana() const { return m_wireMaxMana; }
    // Fallback when the wire max is absent (Live): highest current-mana ever
    // seen. You can't hold more current than your max, so this equals the exact
    // max once the player has regen'd to full (beats the calcMaxMana estimate).
@@ -362,7 +362,9 @@ public:
   QString m_stance;           // EQL active stance name (resolved); empty if none
   QString m_invocation;       // EQL active invocation name (resolved); empty if none
   uint16_t m_observedMaxMana = 0;   // peak current mana = exact max once full
-  uint16_t m_wireMaxMana = 0;       // eql: exact gear+buff max from stat-sync wide form
+  // eql: exact gear+buff max from stat-sync wide form. uint32 not uint16 —
+  // the wide form carries 64-bit values and setMana takes uint32.
+  uint32_t m_wireMaxMana = 0;
   uint32_t m_money = 0;   // carried coin as total copper (eql OP_PlayerProfile)
   uint16_t m_fatigue;
   uint32_t m_enduranceCur;
